API documentation

Documentation for Bramble.jl's public API.

Geometry

Bramble.Interval — Function

Interval(x, y)

Creates an interval set from two scalars x and y.

Fields

x – the lower bound

y – the upper bound

Example

julia> Interval(0.0, 1.0)
CartesianProduct{1,Float64}((0.0,1.0))

Bramble.:× — Function

×(X, Y)

Compute the Cartesian product of two Cartesian products X and Y.

Fields

X – the first Cartesian product

Y – the second Cartesian product

Example

julia> X = CartesianProduct(0.0, 1.0);
	   Y = CartesianProduct(2.0, 3.0);
	   X × Y;
CartesianProduct{2,Float64}(((0.0, 2.0), (0.0, 3.0)), ((1.0, 2.0), (1.0, 3.0)))

Bramble.markers — Function

markers(p...)

Converts pairs of Pair{String, Function} into a Tuple{Marker}.

markers(domain)

Returns the markers associated with a domain.

Bramble.Domain — Type

struct Domain{SetType, MarkersType}

Represents a domain with a set and a set of markers.

Inputs:
set::SetType, the set represented by the domain.
markers::MarkersType, the markers associated with the domain.

Mesh

Bramble.ndofs — Method

ndofs(mesh::MeshType)

Return the number of points in the mesh

Bramble.Mesh — Method

Mesh(domain::Domain{CartesianProduct{1,T}, MarkersType}, npts::Int, unif::Bool) where {T, MarkersType}

Create a 1D mesh based on domain and npts with uniform spacing if unif is true (otherwise, the points are randomly generated).

#Example

julia> mesh = Mesh(Domain(I), 10, true);
I = Interval(0,1);

Bramble.Mesh — Method

Mesh(domain::Domain{CartesianProduct{1,T}, MarkersType}, npts::NTuple{1,Int}, uniform::NTuple{1,Bool}) where {T, MarkersType}

Create a 1D mesh based on domain and npts with uniform spacing if unif is true (otherwise, the points are randomly generated).

Bramble.hₘₐₓ — Method

hₘₐₓ(mesh::Mesh1D)

Compute the maximum of the space step sizes in the mesh.

Bramble.points — Method

points(mesh::Mesh1D, i)

Return the point at index i.

Bramble.points — Method

points(mesh::Mesh1D)

Return a vector of all points in the mesh.

Bramble.Mesh — Method

Mesh(domain, npts, unif)

Create a cartesian nD-mesh from the domain domain. The number of points for each projection mesh are given in the tuple npts. The distribution of points on the submeshes are encoded in the tuple unif. The mesh is defined by the markers in markers, which stores which indices correspond to which marker.

Example

julia> mesh = Mesh(domain, (10, 15), (true, true))
domain = Domain(Interval(0,1) × Interval(4,5))

Bramble.hₘₐₓ — Method

hₘₐₓ(mesh)

Return the maximum diagonal of the mesh.

Bramble.points — Method

points(mesh)

Return the points of the mesh.

Bramble.points — Method

points(mesh, idx)

Return the point at index idx in the mesh.