Forms

Bramble.BilinearForm — Type

struct BilinearForm{TrialType,TestType,F} <: BilinearFormType
	trial_space::TrialType
	test_space::TestType
	form_expr::F
end

Structure to store the data associated with a bilinear form

\[\begin{array}{rcll} a \colon & W_h \times V_h &\longrightarrow &\mathbb{R} \\ & (u,v) &\longmapsto & a(u,v). \end{array}\]

The field form_expr has the expression of the form and the remaining fields store the trial and test spaces $W_h$ and $V_h$.

Bramble.BilinearFormType — Type

BilinearFormType

Abstract type for bilinear forms.

Bramble.LinearForm — Type

struct LinearForm{TestType,F} <: LinearFormType
	test_space::TestType
	form_expr::F
end

Structure to store the data associated with a llinear form

\[\begin{array}{rcll} l \colon & W_h &\longrightarrow &\mathbb{R} \\ & v &\longmapsto & l(v). \end{array}\]

The field form_expr has the expression of the form and the remaining field stores the test space $W_h$.

Bramble.LinearFormType — Type

LinearFormType

Abstract type for linear forms.

Bramble.testspace — Method

testspace(a::BilinearForm)

Returns the test space of a bilinear form.

Bramble.testspace — Method

testspace(a::LinearForm)

Returns the test space of a linear form.

Bramble.trialspace — Method

trialspace(a::BilinearForm)

Returns the trial space of a bilinear form.

Bramble.Constraints — Type

struct Constraints{D,FType} <: ConstraintsType
	markers::NTuple{D,Marker{FType}}
	constraint_type::Symbol
end

Structure to store boundary constraints. The constraint_type should be a symbol identifying the type of conditions (currently, only :dirichlet is supported). The tuple markers stores all the Markers informations related with the constraints.

Bramble.ConstraintsType — Type

ConstraintsType

Abstract type for boundary conditions contraints.

Bramble.__set_diag_one — Method

__set_diag_one(A::AbstractMatrix, npts, rows)

Set the diagonal elements of matrix A to one for the given rows.

Bramble.__set_rows_zero — Method

__set_rows_zero(A::AbstractMatrix, npts, rows)

Set the elements of matrix A to zero for the given rows.

Bramble._apply_dirichlet_bc! — Method

_apply_dirichlet_bc!(A::AbstractMatrix, npts, rows)

Apply Dirichlet boundary conditions to matrix A for the given rows`.

Bramble.apply_dirichlet_bc! — Method

apply_dirichlet_bc!(A, bcs::Constraints, M::MeshType)

Apply Dirichlet boundary conditions to matrix A using the Constraints object bcs and the mesh M. For each index i associated with a Dirichlet boundary condition, we set the i-th row of matrix A to zero and change the diagonal element A[i,i] to 1.

Bramble.apply_dirichlet_bc! — Method

apply_dirichlet_bc!(v::AbstractVector, bcs::Constraints, Ωₕ::MeshType)

Apply Dirichlet boundary conditions to vector v using the Constraints object bcs and the mesh Ωₕ.

Bramble.constraint_type — Method

constraint_type(bcs::Constraints)

Returns the symbol defining the type of boundary condition stored in the Constraints object bcs.

Bramble.symbols — Method

symbols(bcs::Constraints)

Returns an iterator over the symbols of the Markers stored in the Constraints object bcs.