API

Documentation for Bramble.jl's public API.

Geometry

Bramble.cartesianproduct — Function

cartesianproduct(data::NTuple)

Returns a CartesianProduct from a tuple of intervals.

cartesianproduct(x, y)

Returns a 1D CartesianProduct from two scalars x and y, where x and y are, respectively, the lower and upper bounds of the interval.

Example

julia> cartesianproduct(0, 1)
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Bramble.interval — Function

interval(x, y)

Returns a 1D CartesianProduct set from two scalars x and y, where x and y are, respectively, the lower and upper bounds of the interval.

Example

julia> interval(0, 1)
CartesianProduct{1,Float64}((0.0,1.0))

Bramble.:× — Function

×(X::CartesianProduct, Y::CartesianProduct)

Returns the cartesian product of two CartesianProduct X and Y as a CartesianProduct.

Example

julia> X = cartesianproduct(0, 1); Y = cartesianproduct(2, 3);
	   X × Y;
Type: Float64 
 Dim: 2 
 Set: [0.0, 1.0] × [2.0, 3.0]

Bramble.domain — Function

Domain(Ω::CartesianProduct)

Creates a Domain from a CartesianProduct assuming the single Marker "Dirichlet" => x -> zero(eltype(x)).

Example

julia> domain(Interval(0,1))
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Boundary markers: Dirichlet

Domain(Ω::CartesianProduct, markers::MarkersType)

Creates a Domain from a CartesianProduct assuming the single Marker "Dirichlet" => x -> zero(eltype(x)).

Example

julia> m = markers( "Dirichlet" => (x -> x-1), "Neumann" => (x -> x-0) ); Domain(Interval(0,1), m)
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Boundary markers: Dirichlet, Neumann

Bramble.create_markers — Function

create_markers(m::MarkerType...)

Converts several Pair{String,F} ("label" => func) to domain Markers to be passed in the construction of a Domain Ω.

Example

julia> create_markers( "Dirichlet" => (x -> x-1), "Neumann" => (x -> x-0) )

Bramble.markers — Function

markers(Ω::Domain)

Returns a generator with the Markers associated with a Domain Ω.

Bramble.labels — Function

labels(Ω::Domain)

Returns a generator with the labels of the Markers associated with a Domain Ω.

Mesh

Bramble.mesh — Function

mesh(Ω::Domain, npts::Int, unif::Bool)

Returns a Mesh1D based on Domain Ω and npts points with uniform spacing if unif is true (otherwise, the points are randomly generated on the domain).

For future reference, we denote the npts entries of vector pts as

\[x_i, \, i=1,\dots,N.\]

Example

julia> I = interval(0,1); Ωₕ = mesh(domain(I), 10, true)
1D Mesh
nPoints: 10
Markers: Dirichlet

mesh(Ω::Domain, npts::NTuple{D,Int}, unif::NTuple{D,Bool})

Returns a MeshnD ($1 \leq n \leq 3$) from the Domain Ω. The number of points for each coordinate projection mesh are given in the tuple npts. The distribution of points on the submeshes are encoded in the tuple unif. The Domain markers are translated to markers as for Mesh1D.

For future reference, the mesh points are denoted as

2D mesh, with npts = ($N_x$, $N_y$)

\[(x_i,y_j), \, i=1,\dots,N_x, \, j=1,\dots,N_y\]

3D mesh, with npts = ($N_x$, $N_y$, $N_z$)

\[(x_i,y_j,z_l), \, i=1,\dots,N_x, \, j=1,\dots,N_y, \, l=1,\dots,N_z.\]

Example

julia> X = Domain(Interval(0,1) × Interval(4,5)); Ωₕ = Mesh(X, (10, 15), (true, false))
2D Mesh
nPoints: 150
Markers: ["Dirichlet"]

Submeshes:
  x direction | nPoints: 10
  y direction | nPoints: 15

Bramble.hₘₐₓ — Function

hₘₐₓ(Ωₕ::Mesh1D)

Returns the maximum over the space stepsize $h_i$of mesh Ωₕ

\[h_{max} = \max_{i=1,\dots,N} x_i - x_{i-1}.\]

hₘₐₓ(Ωₕ::MeshnD)

Returns the maximum diagonal of mesh Ωₕ

2D mesh

\[\max_{i,j} \Vert (h_{x,i}, h_{y,j}) \Vert_2\]

3D mesh

\[\max_{i,j,l} \Vert (h_{x,i}, h_{y,j}, h_{z,l}) \Vert_2\]

Space