API documentation

Documentation for Bramble.jl's public API.

Geometry

Bramble.Interval — Function

Interval(x, y)

Creates an interval set from two scalars x and y.

Fields

x, the lower bound

y, the upper bound

Example

julia> Interval(0.0, 1.0)
CartesianProduct{1,Float64}((0.0,1.0))

Bramble.:× — Function

×(X, Y)

Compute the cartesian product of two cartesian products X and Y.

Fields

X – the first set

Y – the second set

Example

julia> X = CartesianProduct(0, 1); Y = CartesianProduct(2, 3);
	   X × Y;
Type: Float64 
 Dim: 2 
 Set: [0.0, 1.0] × [2.0, 3.0]

Bramble.CartesianProduct — Type

Creates a cartesian product of D intervals with elements of type T.

Fields

data, a D-tuple containing the intervals defining the coordinate projections as 2-tuples.

Bramble.markers — Function

markers(p...)

Converts pairs of "label" => func to domain markers to be accepted in the Domain constructor.

Example

julia> markers( "Dirichlet" => (x -> x-1), "Neumann" => (x -> x-0) )

markers(domain)

Returns a generator with the markers associated with a domain.

Bramble.labels — Function

labels(domain)

Returns a generator with the labels of the markers associated with a domain.

Bramble.Domain — Type

struct Domain{SetType, MarkersType}
	set::SetType
	markers::MarkersType
end

Structure to represent a domain composed of cartesian set and a set of markers.