API documentation

Documentation for Bramble.jl's public API.

Geometry

Bramble.CartesianProduct — Type

struct CartesianProduct{D,T}
	data::NTuple{D,Tuple{T,T}}
end

Type for storage of cartesian products of D intervals having elements of type T.

Bramble.Interval — Function

Interval(x, y)

Returns a 1D CartesianProduct set from two scalars x and y, where x and y are, respectively, the lower and upper bounds of the interval.

Example

julia> Interval(0, 1)
CartesianProduct{1,Float64}((0.0,1.0))

Bramble.:× — Function

×(X, Y)

Returns the cartesian product of two CartesianProduct X and Y as a CartesianProduct.

Example

julia> X = CartesianProduct(0, 1); Y = CartesianProduct(2, 3);
	   X × Y;
Type: Float64 
 Dim: 2 
 Set: [0.0, 1.0] × [2.0, 3.0]

Bramble.Domain — Type

struct Domain{SetType, MarkersType}
	set::SetType
	markers::MarkersType
end

Structure to represent a domain composed of a CartesianProduct and a set of Markers.

Bramble.Domain — Method

Domain(X)

Creates a Domain from a CartesianProduct assuming the single Marker "Dirichlet" => x -> zero(eltype(x)).

Example

julia> Domain(Interval(0,1))
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Boundary markers: Dirichlet

Bramble.create_markers — Function

create_markers(m)

Converts several Pair{String,F} ("label" => func) to domain Markers to be passed in the construction of a Domain X.

Example

julia> create_markers( "Dirichlet" => (x -> x-1), "Neumann" => (x -> x-0) )

Bramble.markers — Function

markers(X)

Returns a generator with the Markers associated with a Domain X.

Bramble.labels — Function

labels(X)

Returns a generator with the labels of the Markers associated with a Domain X.