Full internal documentation

Documentation for Bramble.jl's functions.

Geometry

Base.eltype — Method

eltype(X)

Returns the element type of a CartesianProduct.

Example

julia> X = CartesianProduct(0, 1); eltype(X)
Float64

Bramble.dim — Method

dim(X)

Returns the topological dimension of a CartesianProduct.

Example

julia> X = CartesianProduct(0, 1); dim(X)
1

Bramble.projection — Method

projection(X, i)

Returns the i-th 1D CartesianProduct of the CartesianProduct X.

Example

julia> X = CartesianProduct(0, 1) × CartesianProduct(4, 5); projection(X, 1)
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Bramble.tails — Method

tails(X, i)

Returns a tuple with the 1D CartesianProduct of the i-th interval of the CartesianProduct X.

Example

julia> X = CartesianProduct(0, 1) × CartesianProduct(4, 5); tails(X,1)
(0.0, 1.0)

Bramble.tails — Method

tails(X)

Returns a tuple of tuples with 1D CartesianProducts that make up the CartesianProduct X.

Example

julia> X = CartesianProduct(0, 1) × CartesianProduct(4, 5); tails(X)
((0.0, 1.0), (4.0, 5.0))

Bramble.DomainBaseType — Type

DomainBaseType

An abstract type for representing domains.

Bramble.Marker — Type

struct Marker{F<:Function}
	label::String
	f::F
end

Structure to implement markers for a portion of a domain. Each Marker is composed of a label and a levelset function that identifies a portion of the domain.

Base.eltype — Method

eltype(X)

Returns the element type of a Domain X.

Example

julia> eltype(Domain(I × I))
Float64

Bramble.dim — Method

dim(X)

Returns the topological dimension of a Domain X.

Example

julia> I = Interval(0.0, 1.0); dim(Domain(I × I))
2

Bramble.marker_funcs — Method

marker_funcs(X)

Returns a generator with the Markers levelset functions associated with a Domain X.

Bramble.projection — Method

projection(X, i)

Returns the CartesianProduct of the i-th projection of the set of the Domain X.

For example, projection(Domain(I × I), 1) will return I.

Bramble.set — Method

set(X)

Returns the CartesianProduct associated with a Domain X.

Mesh

1D Meshes

Bramble.Mesh1D — Type

Mesh1D{T}(markers, indices, pts, npts)

Create a 1D mesh with `npts` points of type `T`. The points that define the mesh are 
stored in `pts` and are identified, following the same order, with the
indices in `indices`. The mesh is defined by the markers in `markers`, 
which stores which indices correspond to which marker.

Bramble.Mesh1D — Method

(mesh::Mesh1D)()

Return the mesh itself.

Base.eltype — Method

eltype(mesh::Mesh1D{T})

Return the type of the points in mesh.

Base.eltype — Method

eltype(mesh::Mesh1D)

Return the type of the points in the mesh

Base.show — Method

show(io::IO, mesh::Mesh1D)

Print a summary of the mesh to io.

Example

julia> mesh
I = Interval(0,1);

Bramble.addmarkers! — Method

addmarkers!(mrks::MeshMarkers{1}, domain::Domain, R::CartesianIndices{1}, pts)

Add markers to the mesh based on the domain.

Bramble.bcindices — Method

bcindices(mesh::Mesh1D)

Compute the boundary indices of the mesh.

Bramble.bcindices — Method

bcindices(R::CartesianIndices{1})

Compute the boundary indices of the mesh.

Bramble.create_mesh1d_basics — Method

create_mesh1d_basics(domain::Domain, npts::Int, unif::Bool)

Create the basic components of a 1D mesh.

Bramble.createpoints! — Method

createpoints!(x::Vector{T}, I::CartesianProduct{1,T}, npts::Int, unif::Bool) where T

Create a vector of points x uniformly or randomly distributed in the interval I.

Bramble.dim — Method

dim(mesh::Mesh1D{T})

Return the topologicaldimension of the domain associated with mesh.

Bramble.dim — Method

dim(mesh::Mesh1D)

Return the topological dimension of the mesh

Bramble.generate_indices — Method

generate_indices(npts::Int)

Generate a CartesianIndices object for a 1D array of length npts.

Bramble.hmean — Method

hmean(mesh::Mesh1D{T}, i)

Compute the mean of the space step size at index i in the mesh.

Bramble.hmeanit — Method

hmeanit(mesh::Mesh1D)

Return an iterator over all mean space step sizes in the mesh.

Bramble.hspace — Method

hspace(mesh::Mesh1D, i)

Compute the space step size at index i in the mesh.

Bramble.hspaceit — Method

hspaceit(mesh::Mesh1D)

Return an iterator over all space step sizes in the mesh.

Bramble.intindices — Method

intindices(mesh::Mesh1D)

Compute the interior indices of the mesh.

Bramble.intindices — Method

intindices(R::CartesianIndices{1})

Compute the interior indices of the mesh.

Bramble.meas_cell — Method

meas_cell(mesh::Mesh1D, idx::CartesianIndex{1})

Compute the measure of the cell at index idx in the mesh.

Bramble.meas_cellit — Method

meas_cellit(mesh::Mesh1D)

Return an iterator over all cell measures in the mesh.

Bramble.npoints — Method

npoints(mesh::Mesh1D)

Return a tuple of the number of points in the mesh.

#Example

julia> npoints(mesh)
I = Interval(0,1);

Bramble.pointsit — Method

pointsit(mesh::Mesh1D)

Return an iterator over all points in the mesh.

Bramble.xmean — Method

xmean(mesh::Mesh1D{T}, i)

Compute the mean of the points at index i in the mesh.

Bramble.xmeanit — Method

xmeanit(mesh::Mesh1D)

Return an iterator over all mean points in the mesh.

nD Meshes, $n>1$

Bramble.MeshnD — Type

MeshnD{n,T}(markers, indices, pts, npts)

Create a cartesian nD-mesh with `npts` points. For efficiency,
the mesh points are not stored. Instead, we store the points
of the 1D meshes that make up the nD mesh. To connect both nD and 
1D meshes, we use the indices in `indices`. The mesh is defined 
by the markers in `markers`, which stores which indices 
correspond to which marker.

Bramble.MeshnD — Method

mesh(i)

Return the i-th submesh.

Base.eltype — Method

eltype(mesh)

Return the eltype of the mesh.

Base.eltype — Method

eltype(mesh::MeshnD)

Return the type of the points in the mesh

Base.show — Method

show(io, mesh)

Print a description of the mesh to the IO stream io.

Bramble.addmarkers! — Method

addmarkers!(markers, domain, meshes)

Add the markers of the domain to the markers of the mesh, using the submeshes.

Bramble.bcindices — Method

bcindices(mesh)

Return the boundary indices of the mesh.

Bramble.dim — Method

dim(mesh)

Return the dimension of the mesh.

Bramble.dim — Method

dim(mesh::MeshnD)

Return the topological dimension of the mesh

Bramble.generate_indices — Method

generate_indices(nPoints)

Return the indices of a mesh with nPoints[i] in each direction.

Bramble.hmean — Method

hmean(mesh, idx)

Compute the mean of the space step sizes at index idx in the mesh.

Bramble.hmeanit — Method

hmeanit(mesh)

Return an iterator over all mean space step sizes in the mesh.

Bramble.hspace — Method

hspace(mesh, idx)

Compute the space step size at index idx in the mesh.

Bramble.hspaceit — Method

hspaceit(mesh)

Return an iterator over all space step sizes in the mesh.

Bramble.intindices — Method

intindices(mesh)

Return the interior indices of the mesh.

Bramble.is_boundary_index — Method

is_boundary_index(idx, R)

Return true if the index idx is a boundary index of the mesh with indices stored in R.

Bramble.meas_cell — Method

meas_cell(mesh, idx)

Return the measure of the cell at the index idx.

Bramble.npoints — Method

npoints(mesh)

Return the number of points of the mesh.

Bramble.pointsit — Method

pointsit(mesh)

Return an iterator over all points in the mesh.

Bramble.xmean — Method

xmean(mesh, idx)

Compute the mean of the points at index idx in the mesh.

Bramble.xmeanit — Method

xmeanit(mesh)

Return an iterator over all mean points in the mesh.