Internal

Documentation for Bramble.jl's functions.

Geometry

Bramble.CartesianProduct — Type

struct CartesianProduct{D,T}
	data::NTuple{D,Tuple{T,T}}
end

Type for storage of cartesian products of D intervals having elements of type T.

Base.eltype — Method

eltype(X::CartesianProduct)

Returns the element type of a CartesianProduct.

Example

julia> X = cartesianproduct(0, 1); eltype(X)
Float64

Bramble.dim — Method

dim(X::CartesianProduct)

Returns the topological dimension of a CartesianProduct.

Example

julia> X = cartesianproduct(0, 1); dim(X)
1

Bramble.projection — Method

projection(X::CartesianProduct, i)

Returns the i-th 1D CartesianProduct of the CartesianProduct X.

Example

julia> X = cartesianproduct(0, 1) × cartesianproduct(4, 5); projection(X, 1)
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Bramble.tails — Method

tails(X::CartesianProduct, i)

Returns a tuple with the 1D CartesianProduct of the i-th interval of the CartesianProduct X.

Example

julia> X = cartesianproduct(0, 1) × cartesianproduct(4, 5); tails(X,1)
(0.0, 1.0)

Bramble.tails — Method

tails(X::CartesianProduct)

Returns a tuple of tuples with 1D CartesianProducts that make up the CartesianProduct X.

Example

julia> X = cartesianproduct(0, 1) × cartesianproduct(4, 5); tails(X)
((0.0, 1.0), (4.0, 5.0))

Bramble.Domain — Type

struct Domain{SetType, MarkersType}
	set::SetType
	markers::MarkersType
end

Structure to represent a domain composed of a CartesianProduct and a set of Markers.

Bramble.DomainBaseType — Type

DomainBaseType

An abstract type for representing domains.

Bramble.Marker — Type

struct Marker{F<:Function}
	label::String
	f::F
end

Structure to implement markers for a portion of a domain. Each Marker is composed of a label and a levelset function that identifies a portion of the domain.

Base.eltype — Method

eltype(Ω::Domain)

Returns the element type of a Domain Ω.

Example

julia> eltype(Domain(I × I))
Float64

Bramble.dim — Method

dim(Ω::DomainBaseType)

Returns the topological dimension of a Domain Ω.

Example

julia> I = Interval(0.0, 1.0); dim(Domain(I × I))
2

Bramble.marker_funcs — Method

marker_funcs(Ω::Domain)

Returns a generator with the Markers levelset functions associated with a Domain Ω.

Bramble.projection — Method

projection(Ω::Domain, i::Int)

Returns the CartesianProduct of the i-th projection of the set of the Domain Ω.

For example, projection(Domain(I × I), 1) will return I.

Bramble.set — Method

set(Ω::Domain)

Returns the CartesianProduct associated with a Domain Ω.

Mesh

1D Meshes

Bramble.Mesh1D — Type

struct Mesh1D{T} <: MeshType{1}
	markers::MeshMarkers{1}
	indices::CartesianIndices{1,Tuple{Base.OneTo{Int}}}
	pts::Vector{T}
	npts::Int
end

Structure to create a 1D mesh with npts points of type T. The points that define the mesh are stored in pts and are identified, following the same order, with the indices in indices. The variable markers stores, for each Domain marker, the indices satisfying $f(x)=0$, where $f$is the marker function.

For future reference, the npts entries of vector pts are

\[x_i, \, i=1,\dots,N.\]

Bramble.addmarkers! — Method

addmarkers!(markerList::MeshMarkers{1}, Ω::Domain, R::CartesianIndices{1}, pts)

For each Domain marker, stores in markerList the indices of the points that satisfy $f(x)=0$, where $f$ is the corresponding levelset function associated with the marker.

Bramble.bcindices — Method

bcindices(Ωₕ::Mesh1D)

Returns the indices of the boundary points of mesh Ωₕ.

Bramble.cell_measure — Method

cell_measure(Ωₕ::Mesh1D, i::CartesianIndex)

Returns the measure of the cell $\square_{i} = [x_i - \frac{h_{i}}{2}, x_i + \frac{h_{i+1}}{2}]$ at CartesianIndex i in mesh Ωₕ, which is given by $h_{i+1/2}$.

Bramble.cell_measure_iterator — Method

cell_measure_iterator(Ωₕ::Mesh1D)

Returns an iterator over $h_{i+1/2}, \, i=1,\dots,N$ in mesh Ωₕ.

Bramble.create_mesh1d_basics — Method

create_mesh1d_basics(Ω::Domain, npts::Int, unif::Bool)

Creates the basic components of a 1D mesh, given a Domain Ω, the number of points npts and a boolean unif. The points are equally spaced if unif is true (otherwise, the points are randomly generated on the domain).

Bramble.createpoints! — Method

createpoints!(x::Vector, I::CartesianProduct{1}, unif::Bool)

Overrides the components of vector x with uniformly (unif = true) or randomly distributed (unif = false) points in the interval I.

Bramble.generate_indices — Method

generate_indices(npts::Int)

Returns a CartesianIndices object for a vector of length npts.

Bramble.half_points — Method

function half_points(Ωₕ::Mesh1D, i)

Returns the average of two neighboring, $x_{i+1/2}$, points in mesh Ωₕ, at index i

\[x_{i+1/2} = x_i + \frac{h_{i+1}}{2}, \, i=1,\dots,N-1,\]

$x_{N+1/2} = x_{N}$ and $x_{1/2} = x_1$.

Bramble.half_points_iterator — Method

half_points_iterator(Ωₕ::Mesh1D)

Returns an iterator over all points $x_{i+1/2}, \, i=1,\dots,N$ in mesh Ωₕ.

Bramble.half_spacing — Method

half_spacing(Ωₕ::Mesh1D, i)

Returns the indexwise average of the space stepsize, $h_{i+1/2}$, at index i in mesh Ωₕ

\[h_{i+1/2} = \frac{h_i + h_{i+1}}{2}, \, i=1,\dots,N-1\]

where $h_{N+1/2} = \frac{h_{N}}{2}$ and $h_{1/2} = \frac{h_1}{2}$.

Bramble.half_spacing_iterator — Method

half_spacing_iterator(Ωₕ::Mesh1D)

Returns an iterator over all indexwise average of the space stepsizes $h_{i+1/2}, \, i=1,\dots,N$ in mesh Ωₕ.

Bramble.intindices — Method

intindices(Ωₕ::Mesh1D)

Returns the indices of the interior points of mesh Ωₕ.

Bramble.npoints — Method

npoints(Ωₕ::Mesh1D)

Returns a 1-tuple of the number of points $x_i$ in Ωₕ.

Example

julia> Ωₕ = Mesh(Domain(Interval(0,1)), 10, true); npoints(Ωₕ)
(10,)

Bramble.point — Method

point(Ωₕ::Mesh1D, i)

Returns the i-th point of Ωₕ, $x_i$.

Bramble.points — Method

points(Ωₕ::Mesh1D)

Returns a vector with all the points $x_i, \, i=1,\dots,N$ in Ωₕ.

Bramble.points_iterator — Method

points_iterator(Ωₕ::Mesh1D)

Returns an iterator over all points $x_i, \, i=1,\dots,N$ in mesh Ωₕ.

Bramble.spacing — Method

spacing(Ωₕ::Mesh1D, i)

Returns the space stepsize, $h_i$ at index i in mesh Ωₕ

\[h_i = x_i - x_{i-1}, \, i=2,\dots,N\]

where $h_1 = x_2 - x_1$.

Bramble.spacing_iterator — Method

spacing_iterator(Ωₕ::Mesh1D)

Returns an iterator over all space step sizes $h_i, \, i=1,\dots,N$ in mesh Ωₕ.

nD Meshes, $n=2,3$

Bramble.MeshnD — Type

struct MeshnD{n,T} <: MeshType{n}
	markers::MeshMarkers{n}
	indices::CartesianIndices{n,NTuple{n,UnitRange{Int}}}
	npts::Int
	submeshes::NTuple{n,Mesh1D{T}}
end

Type to store a cartesian nD-mesh ($2 \leq n \leq 3$) with prod(npts) points of type T. For efficiency, the mesh points are not stored. Instead, we store the points of the 1D meshes that make up the nD mesh. To connect both nD and 1D meshes, we use the indices in indices. The Domain markers are translated to markers as for Mesh1D.

Bramble.MeshnD — Method

(Ωₕ::MeshnD)(i)

Returns the i-th submesh of Ωₕ.

Bramble.addmarkers! — Method

addmarkers!(mrks::MeshMarkers, Ω::Domain, submeshes::NTuple{D,Mesh1D})

Adds the markers of Domain to the markers of the mesh, using the submeshes` in each coordinate direction.

Bramble.boundary_indices — Method

boundary_indices(Ωₕ::MeshnD)

Returns the boundary indices of mesh $Ωₕ$.

Bramble.cell_measure — Method

cell_measure(Ωₕ::MeshnD, idx::CartesianIndex)

Returns the measure of the cell $\square_{idx}$ centered at the index idx

2D mesh, $\square_{i,j} = [x_i - \frac{h_{x,i}}{2}, x_i + \frac{h_{x,i+1}}{2}] \times [y_j - \frac{h_{y,j}}{2}, y_j + \frac{h_{y,j+1}}{2}]$ is

\[h_{x,i+1/2} h_{y,j+1/2}\]

where idx = $(i,j)$,

3D mesh, $\square_{i,j,l} = [x_i - \frac{h_{x,i}}{2}, x_i + \frac{h_{x,i+1}}{2}] \times [y_j - \frac{h_{y,j}}{2}, y_j + \frac{h_{y,j+1}}{2}] \times [z_l - \frac{h_{z,l}}{2}, z_l + \frac{h_{z,l+1}}{2}]$ is

\[h_{x,i+1/2} h_{y,j+1/2} h_{z,l+1/2}\]

where idx = $(i,j,l)$.

Bramble.generate_indices — Method

generate_indices(nPoints::NTuple)

Returns the CartesianIndices indices of a mesh with nPoints[i] in each direction.

Bramble.half_points — Method

half_points(Ωₕ::MeshnD, idx::NTuple)

Returns a tuple with the half_points, for each submesh, of the points at index idx

2D mesh, with idx = $(i,j)$

\[(x_{i+1/2}, y_{j+1/2})\]

3D mesh, with idx = $(i,j,l)$

\[(x_{i+1/2}, y_{j+1/2}, z_{l+1/2}).\]

Bramble.half_points_iterator — Method

half_points_iterator(Ωₕ::MeshnD)

Returns an iterator over all half_points points in mesh Ωₕ

2D mesh

\[(x_{i+1/2}, y_{j+1/2}), i = 1,...,N_{x}, j = 1,...,N_{y}\]

3D mesh

\[(x_{i+1/2}, y_{j+1/2}, z_{l+1/2}), i = 1,...,N_{x}, j = 1,...,N_{y}, l = 1,...,N_{z}.\]

Bramble.half_spacing — Method

half_spacing(Ωₕ::MeshnD, idx::NTuple)

Returns a tuple with the half_spacing, for each submesh, at index idx

2D mesh, with idx = $(i,j)$

\[(h_{x,i+1/2}, h_{y,j+1/2})\]

3D mesh, with idx = $(i,j,l)$

\[(h_{x,i+1/2}, h_{y,j+1/2}, h_{z,l+1/2})\]

Bramble.half_spacing_iterator — Method

half_spacing_iterator(Ωₕ::MeshnD)

Returns an iterator over all mean space step sizes half_spacing in mesh Ωₕ

2D mesh

\[(h_{x,i+1/2}, h_{y,j+1/2}), i = 1,...,N_{x}, j = 1,...,N_{y}\]

3D mesh

\[(h_{x,i+1/2}, h_{y,j+1/2}, h_{z,l+1/2}), i = 1,...,N_{x}, j = 1,...,N_{y}, l = 1,...,N_{z}.\]

Bramble.interior_indices — Method

interior_indices(R::CartesianIndices)

Returns the interior indices of mesh $Ωₕ$.

Bramble.is_boundary_index — Method

is_boundary_index(idx::CartesianIndex, R::CartesianIndices)

Returns true if the index idx is a boundary index of the mesh with indices stored in R.

Bramble.npoints — Method

npoints(Ωₕ::MeshnD)

Returns a tuple with the number of points of mesh Ωₕ, in each coordinate direction.

Bramble.point — Method

point(Ωₕ::MeshnD, idx::NTuple)

Returns the point at index idx of Ωₕ as a tuple

2D mesh, with idx = $(i,j)$

\[(x_i, y_j)\]

3D mesh, with idx = $(i,j,l)$

\[(x_i, y_j, z_l).\]

Bramble.points — Method

points(Ωₕ::MeshnD)

Returns a tuple with the points of Ωₕ

2D mesh, with npts = ($N_x$, $N_y$)

\[([x_i]_{i=1}^{N_x}, [y_j]_{j=1}^{N_y})\]

3D mesh, with npts = ($N_x$, $N_y$, $N_z$)

\[([x_i]_{i=1}^{N_x}, [y_j]_{j=1}^{N_y}, [z_l]_{l=1}^{N_z}).\]

Bramble.points_iterator — Method

points_iterator(Ωₕ::MeshnD)

Returns an iterator over all points in mesh Ωₕ

2D mesh

\[(x_i, y_j), i = 1,...,N_{x}, j = 1,...,N_{y}\]

3D mesh

\[(x_i, y_j, z_l), i = 1,...,N_{x}, j = 1,...,N_{y}, l = 1,...,N_{z}.\]

Bramble.spacing — Method

spacing(Ωₕ::MeshnD, idx::NTuple)

Returns a tuple with the spacing, for each submesh, at index idx

2D mesh, with idx = $(i,j)$

\[(h_{x,i}, h_{y,j}) := (x_i - x_{i-1}, y_j - y_{j-1})\]

3D mesh, with idx = $(i,j,l)$

\[(h_{x,i}, h_{y,j}, h_{z,l}) := (x_i - x_{i-1}, y_j - y_{j-1}, z_l - z_{l-1})\]

Bramble.spacing_iterator — Method

spacing_iterator(Ωₕ::MeshnD{D})

Returns an iterator over all space step sizes spacing in mesh Ωₕ

2D mesh

\[(h_{x,i}, h_{y,j}), i = 1,...,N_{x}, j = 1,...,N_{y}\]

3D mesh

\[(h_{x,i}, h_{y,j}, h_{z,l}), i = 1,...,N_{x}, j = 1,...,N_{y}, l = 1,...,N_{z}.\]