API documentation

Documentation for Bramble.jl's public API.

Contents

Contents

Index

Geometry

Mesh

Index

Bramble.Domain

Bramble.Domain

Bramble.Mesh1D

Bramble.Mesh1D

Bramble.MeshnD

Bramble.MeshnD

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.show

Base.show

Bramble.:×

Bramble.Interval

Bramble.Mesh

Bramble.Mesh

Bramble.Mesh

Bramble.addmarkers!

Bramble.addmarkers!

Bramble.bcindices

Bramble.bcindices

Bramble.bcindices

Bramble.create_mesh1d_basics

Bramble.createpoints!

Bramble.dim

Bramble.dim

Bramble.dim

Bramble.dim

Bramble.dim

Bramble.generate_indices

Bramble.generate_indices

Bramble.hmean

Bramble.hmean

Bramble.hmeanit

Bramble.hmeanit

Bramble.hspace

Bramble.hspace

Bramble.hspaceit

Bramble.hspaceit

Bramble.hₘₐₓ

Bramble.hₘₐₓ

Bramble.indices

Bramble.intindices

Bramble.intindices

Bramble.intindices

Bramble.is_boundary_index

Bramble.marker

Bramble.markers

Bramble.markers

Bramble.meas_cell

Bramble.meas_cell

Bramble.meas_cellit

Bramble.ndofs

Bramble.npoints

Bramble.npoints

Bramble.points

Bramble.points

Bramble.points

Bramble.points

Bramble.pointsit

Bramble.pointsit

Bramble.xmean

Bramble.xmean

Bramble.xmeanit

Bramble.xmeanit

Geometry

Sets

Bramble.:× — Method

×(X, Y)

Compute the Cartesian product of two Cartesian products X and Y.

Fields

X – the first Cartesian product

Y – the second Cartesian product

Example

julia> X = CartesianProduct(0.0, 1.0);
	   Y = CartesianProduct(2.0, 3.0);
	   X × Y;
CartesianProduct{2,Float64}(((0.0, 2.0), (0.0, 3.0)), ((1.0, 2.0), (1.0, 3.0)))

Bramble.Interval — Method

Interval(x, y)

Creates an interval set from two scalars x and y.

Fields

x – the lower bound

y – the upper bound

Example

julia> Interval(0.0, 1.0)
CartesianProduct{1,Float64}((0.0,1.0))

Domains

Bramble.Domain — Type

struct Domain{SetType, MarkersType}

Represents a domain with a set and a set of markers.

Inputs:
set::SetType, the set represented by the domain.
markers::MarkersType, the markers associated with the domain.

Bramble.Domain — Method

The Domain type represents a domain with a set and a set of markers.

The default marker for a domain is a "Dirichlet" marker with a function f(x) = 0.

Bramble.markers — Method

markers(domain)

Returns the markers associated with a domain.

Bramble.markers — Method

markers(p...)

Converts pairs of Pair{String, Function} into a Tuple{Marker}.

Mesh

Base.eltype — Method

eltype(mesh::MeshType)

Return the type of the points in the mesh

Bramble.dim — Method

dim(mesh::MeshType)

Return the dimension of the mesh

Bramble.indices — Method

indices(mesh::MeshType)

Return the indices of the points in the mesh

Bramble.marker — Method

marker(mesh::MeshType, m)

Return the indices of the points in the mesh that have the marker m

Bramble.ndofs — Method

ndofs(mesh::MeshType)

Return the number of points in the mesh

Bramble.Mesh1D — Type

Mesh1D{T}(markers, indices, pts, npts)

Create a 1D mesh with `npts` points of type `T`. The points that define the mesh are 
stored in `pts` and are identified, following the same order, with the
indices in `indices`. The mesh is defined by the markers in `markers`, 
which stores which indices correspond to which marker.

Bramble.Mesh1D — Method

(mesh::Mesh1D)()

Return the mesh itself.

Base.eltype — Method

eltype(mesh::Mesh1D{T})

Return the type of the points in mesh.

Base.eltype — Method

eltype(mesh::Mesh1D)

Return the type of the points in the mesh

Base.show — Method

show(io::IO, mesh::Mesh1D)

Print a summary of the mesh to io.

Example

julia> mesh
I = Interval(0,1);

Bramble.Mesh — Method

Mesh(domain::Domain{CartesianProduct{1,T}, MarkersType}, npts::Int, unif::Bool) where {T, MarkersType}

Create a 1D mesh based on domain and npts with uniform spacing if unif is true (otherwise, the points are randomly generated).

#Example

julia> mesh = Mesh(Domain(I), 10, true);
I = Interval(0,1);

Bramble.Mesh — Method

Mesh(domain::Domain{CartesianProduct{1,T}, MarkersType}, npts::NTuple{1,Int}, uniform::NTuple{1,Bool}) where {T, MarkersType}

Create a 1D mesh based on domain and npts with uniform spacing if unif is true (otherwise, the points are randomly generated).

Bramble.addmarkers! — Method

addmarkers!(mrks::MeshMarkers{1}, domain::Domain, R::CartesianIndices{1}, pts)

Add markers to the mesh based on the domain.

Bramble.bcindices — Method

bcindices(mesh::Mesh1D)

Compute the boundary indices of the mesh.

Bramble.bcindices — Method

bcindices(R::CartesianIndices{1})

Compute the boundary indices of the mesh.

Bramble.create_mesh1d_basics — Method

create_mesh1d_basics(domain::Domain, npts::Int, unif::Bool)

Create the basic components of a 1D mesh.

Bramble.createpoints! — Method

createpoints!(x::Vector{T}, I::CartesianProduct{1,T}, npts::Int, unif::Bool) where T

Create a vector of points x uniformly or randomly distributed in the interval I.

Bramble.dim — Method

dim(mesh::Mesh1D{T})

Return the topologicaldimension of the domain associated with mesh.

Bramble.dim — Method

dim(mesh::Mesh1D)

Return the topological dimension of the mesh

Bramble.generate_indices — Method

generate_indices(npts::Int)

Generate a CartesianIndices object for a 1D array of length npts.

Bramble.hmean — Method

hmean(mesh::Mesh1D{T}, i)

Compute the mean of the space step size at index i in the mesh.

Bramble.hmeanit — Method

hmeanit(mesh::Mesh1D)

Return an iterator over all mean space step sizes in the mesh.

Bramble.hspace — Method

hspace(mesh::Mesh1D, i)

Compute the space step size at index i in the mesh.

Bramble.hspaceit — Method

hspaceit(mesh::Mesh1D)

Return an iterator over all space step sizes in the mesh.

Bramble.hₘₐₓ — Method

hₘₐₓ(mesh::Mesh1D)

Compute the maximum of the space step sizes in the mesh.

Bramble.intindices — Method

intindices(mesh::Mesh1D)

Compute the interior indices of the mesh.

Bramble.intindices — Method

intindices(R::CartesianIndices{1})

Compute the interior indices of the mesh.

Bramble.meas_cell — Method

meas_cell(mesh::Mesh1D, idx::CartesianIndex{1})

Compute the measure of the cell at index idx in the mesh.

Bramble.meas_cellit — Method

meas_cellit(mesh::Mesh1D)

Return an iterator over all cell measures in the mesh.

Bramble.npoints — Method

npoints(mesh::Mesh1D)

Return a tuple of the number of points in the mesh.

#Example

julia> npoints(mesh)
I = Interval(0,1);

Bramble.points — Method

points(mesh::Mesh1D, i)

Return the point at index i.

Bramble.points — Method

points(mesh::Mesh1D)

Return a vector of all points in the mesh.

Bramble.pointsit — Method

pointsit(mesh::Mesh1D)

Return an iterator over all points in the mesh.

Bramble.xmean — Method

xmean(mesh::Mesh1D{T}, i)

Compute the mean of the points at index i in the mesh.

Bramble.xmeanit — Method

xmeanit(mesh::Mesh1D)

Return an iterator over all mean points in the mesh.

Bramble.MeshnD — Type

MeshnD{n,T}(markers, indices, pts, npts)

Create a cartesian nD-mesh with `npts` points. For efficiency,
the mesh points are not stored. Instead, we store the points
of the 1D meshes that make up the nD mesh. To connect both nD and 
1D meshes, we use the indices in `indices`. The mesh is defined 
by the markers in `markers`, which stores which indices 
correspond to which marker.

Bramble.MeshnD — Method

mesh(i)

Return the i-th submesh.

Base.eltype — Method

eltype(mesh)

Return the eltype of the mesh.

Base.eltype — Method

eltype(mesh::MeshnD)

Return the type of the points in the mesh

Base.show — Method

show(io, mesh)

Print a description of the mesh to the IO stream io.

Bramble.Mesh — Method

Mesh(domain, npts, unif)

Create a cartesian nD-mesh from the domain domain. The number of points for each projection mesh are given in the tuple npts. The distribution of points on the submeshes are encoded in the tuple unif. The mesh is defined by the markers in markers, which stores which indices correspond to which marker.

Example

julia> mesh = Mesh(domain, (10, 15), (true, true))
domain = Domain(Interval(0,1) × Interval(4,5))

Bramble.addmarkers! — Method

addmarkers!(markers, domain, meshes)

Add the markers of the domain to the markers of the mesh, using the submeshes.

Bramble.bcindices — Method

bcindices(mesh)

Return the boundary indices of the mesh.

Bramble.dim — Method

dim(mesh)

Return the dimension of the mesh.

Bramble.dim — Method

dim(mesh::MeshnD)

Return the topological dimension of the mesh

Bramble.generate_indices — Method

generate_indices(nPoints)

Return the indices of a mesh with nPoints[i] in each direction.

Bramble.hmean — Method

hmean(mesh, idx)

Compute the mean of the space step sizes at index idx in the mesh.

Bramble.hmeanit — Method

hmeanit(mesh)

Return an iterator over all mean space step sizes in the mesh.

Bramble.hspace — Method

hspace(mesh, idx)

Compute the space step size at index idx in the mesh.

Bramble.hspaceit — Method

hspaceit(mesh)

Return an iterator over all space step sizes in the mesh.

Bramble.hₘₐₓ — Method

hₘₐₓ(mesh)

Return the maximum diagonal of the mesh.

Bramble.intindices — Method

intindices(mesh)

Return the interior indices of the mesh.

Bramble.is_boundary_index — Method

is_boundary_index(idx, R)

Return true if the index idx is a boundary index of the mesh with indices stored in R.

Bramble.meas_cell — Method

meas_cell(mesh, idx)

Return the measure of the cell at the index idx.

Bramble.npoints — Method

npoints(mesh)

Return the number of points of the mesh.

Bramble.points — Method

points(mesh)

Return the points of the mesh.

Bramble.points — Method

points(mesh, idx)

Return the point at index idx in the mesh.

Bramble.pointsit — Method

pointsit(mesh)

Return an iterator over all points in the mesh.

Bramble.xmean — Method

xmean(mesh, idx)

Compute the mean of the points at index idx in the mesh.

Bramble.xmeanit — Method

xmeanit(mesh)

Return an iterator over all mean points in the mesh.