Geometry

Bramble.BrambleBareFunction — Type

struct BrambleBareFunction{D,T,has_cart} 
	f_tuple::FunctionWrapper{T, Tuple{NTuple{D,T}}}
	f_cartesian::FunctionWrapper{T, Tuple{CartesianIndex{D}}}
end

Structure to wrap around functions to make them more type agnostic. It uses FunctionWrappers to provide functions calculated on Tuples and CartesianIndices. The type arguments are D, the dimension of the underlying domain in which the function is defined, f_tuple, a version of the function appliable to Tuples and f_cartesian, a version of the function appliable to CartesianIndices (useful when dealing with meshes)

Bramble.CartesianProduct — Type

struct CartesianProduct{D,T}
	data::NTuple{D,Tuple{T,T}}
end

Type for storage of cartesian products of D intervals having elements of type T.

Bramble.CartesianProduct — Method

(X::CartesianProduct)(i)

Returns the i-th interval in the CartesianProduct.

Base.eltype — Method

eltype(X::CartesianProduct)
eltype(::Type{<:CartesianProduct})

Returns the element type of a CartesianProduct.

Example

julia> X = cartesianproduct(0, 1); eltype(X)
Float64

Bramble.dim — Method

dim(X::CartesianProduct)
dim(::Type{<:CartesianProduct})

Returns the topological dimension of a CartesianProduct.

Example

julia> X = cartesianproduct(0, 1); dim(X)
1

julia> Y = cartesianproduct(((0,1), (4,5))); dim(Y)
2

Bramble.projection — Method

projection(X::CartesianProduct, i)

Returns the i-th interval in CartesianProduct X as a new 1-dimensional CartesianProduct.

Example

julia> X = cartesianproduct(0, 1) × cartesianproduct(4, 5); projection(X, 1)
Type: Float64 
 Dim: 1 
 Set: [0.0, 1.0]

Bramble.tails — Method

tails(X::CartesianProduct, i)
tails(X::CartesianProduct{D})

Returns i-th interval in CartesianProduct X as a Tuple. It can also be called on X, returning aD-tuple with all intervals defining [CartesianProduct](@ref)X`.

Example

julia> X = cartesianproduct(0, 1) × cartesianproduct(4, 5); tails(X,1)
(0.0, 1.0)

julia> X = cartesianproduct(0, 1) × cartesianproduct(4, 5); tails(X)
((0.0, 1.0), (4.0, 5.0))

Bramble.Domain — Type

struct Domain{SetType, MarkersType}
	set::SetType
	markers::MarkersType
end

Structure to represent a domain composed of a CartesianProduct and a set of Markers.

Bramble.Marker — Type

struct Marker{F}
	label::Symbol
	f::F
end

Structure to implement markers for a portion of a domain or even boundary conditions. Each Marker is composed of a symbol and a BrambleBareFunction.

Base.eltype — Method

eltype(Ω::Domain)
eltype(::Type{<:Domain{SetType}})

Returns the type of the bounds defining Domain Ω.

Example

julia> I = interval(0.0, 1.0); eltype(domain(I × I))
Float64

Bramble.dim — Method

dim(Ω::Domain)
dim(::Type{<:Domain})

Returns the topological dimension of the Domain Ω.

Example

julia> I = interval(0.0, 1.0); dim(domain(I × I))
2

Bramble.marker_funcs — Method

marker_funcs(Ω::Domain)

Returns a generator with the Marker's BrambleBareFunctions associated with Domain Ω.

Bramble.projection — Method

projection(Ω::Domain, i)

Returns the i-th 1-dimensional CartesianProduct of the set associated with Domain Ω.

For example, projection(domain(I × I), 1) will return I.

Bramble.set — Method

set(Ω::Domain)

Returns the CartesianProduct associated with the Domain Ω.